Java 1.5 seems to complain about assertEquals(Object, Object), which can be solved by forcing a assertEquals(native, native), e.g. assertEquals(int, int)
git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/cdk/trunk@11063 eb4e18e3-b210-0410-a6ab-dec725e4b171

+++ src/test/org/openscience/cdk/AssociationTest.java
@@ -38,13 +38,13 @@ public class AssociationTest extends NewCDKTestCase {
 
     @Test public void testAssociation() {
         Association association = new Association();
-        Assert.assertEquals(0, association.getElectronCount());
+        Assert.assertEquals(0, association.getElectronCount().intValue());
         Assert.assertEquals(0, association.getAtomCount());
     }
     
     @Test public void testAssociation_IAtom_IAtom() {
         Association association = new Association(new Atom("C"), new Atom("C"));
-        Assert.assertEquals(0, association.getElectronCount());
+        Assert.assertEquals(0, association.getElectronCount().intValue());
         Assert.assertEquals(2, association.getAtomCount());
     }
